Create RebootReminder.ps1
In this commit, I have added a new script named RebootReminder.ps1. This PowerShell script is designed to remind users to reboot their system if it hasn't been rebooted for a specified number of days.

The script sends a balloon notification to the user reminding them to reboot their computer. If the computer isn't rebooted within a specified timeframe (default is 5 hours), the script will enforce a system reboot.

Script Parameters:

DaysLimit: Mandatory parameter. The number of days to check for the last reboot.
HoursLimit: Optional parameter. The number of hours before enforcing a reboot (default is 5 hours).
LogPath: Optional parameter. The path for the log file (default is C:\temp\RebootLog.log).
WorkStart: Optional parameter. The start of the workday in 24-hour format (default is 8).
WorkEnd: Optional parameter. The end of the workday in 24-hour format (default is 17).
This script is intended to be run using the Windows Task Scheduler. Users are advised to schedule this script to run once per day to check for the last reboot time.